Split Codex Connector review policy
