+++ src/codex-connector-review-policy.test.ts
@@ -1,0 +1,127 @@
+import assert from "node:assert/strict";
+import test from "node:test";
+import {
+  buildCodexConnectorP2P3PolicyDiagnostic,
+  buildCodexConnectorPolicyBlockDiagnostic,
+  clusterConfiguredBotReviewThreads,
+  codexConnectorMustFixReviewThreads,
+  codexConnectorStaleReviewCommitThreads,
+  evaluateCodexConnectorConvergencePolicy,
+} from "./codex-connector-review-policy";
+import type { GitHubPullRequest, ReviewThread } from "./core/types";
+import { createConfig, createReviewThread } from "./turn-execution-test-helpers";
+
+function codexThread(overrides: Partial<ReviewThread> & { body: string }): ReviewThread {
+  return createReviewThread({
+    ...overrides,
+    comments: {
+      nodes: [
+        {
+          id: "comment-1",
+          body: overrides.body,
+          createdAt: "2026-03-11T00:00:00Z",
+          url: "https://example.test/pr/44#discussion_r1",
+          author: {
+            login: "chatgpt-codex-connector",
+            typeName: "Bot",
+          },
+        },
+      ],
+    },
+  });
+}
+
+test("Codex Connector policy classifies must-fix, nitpick, and stale review commit findings", () => {
+  const config = createConfig({ reviewBotLogins: ["chatgpt-codex-connector"] });
+  const p2Thread = codexThread({
+    id: "thread-p2",
+    body: "P2: Preserve failed restore cleanup as a blocking verification failure.",
+  });
+  const p3NitpickThread = codexThread({
+    id: "thread-p3-softened",
+    body: "P3: Nitpick: prefer a shorter helper name for readability.",
+  });
+  const p3RiskThread = codexThread({
+    id: "thread-p3-escalated",
+    body: "P3: This cleanup can cause a regression in the restore failure path.",
+  });
+  const pr: Pick<
+    GitHubPullRequest,
+    "headRefOid" | "configuredBotCurrentHeadObservedAt" | "configuredBotLatestReviewedCommitSha"
+  > = {
+    headRefOid: "head-new",
+    configuredBotCurrentHeadObservedAt: null,
+    configuredBotLatestReviewedCommitSha: "head-old",
+  };
+
+  assert.deepEqual(codexConnectorMustFixReviewThreads([p2Thread, p3NitpickThread, p3RiskThread]), [
+    p2Thread,
+    p3RiskThread,
+  ]);
+  assert.deepEqual(codexConnectorStaleReviewCommitThreads(pr, [p2Thread]), [p2Thread]);
+  assert.deepEqual(buildCodexConnectorP2P3PolicyDiagnostic(config, [p2Thread, p3NitpickThread, p3RiskThread]), {
+    p2Actionable: 1,
+    p3Softened: 1,
+    p3Escalated: 1,
+  });
+});
+
+test("Codex Connector policy diagnostics and convergence stay focused in the policy module", () => {
+  const config = createConfig({ reviewBotLogins: ["chatgpt-codex-connector"] });
+  const p1Thread = codexThread({
+    id: "thread-p1",
+    path: "src/policy.ts",
+    line: 12,
+    body: "P1: Tighten the diagnostic wording before merge.",
+  });
+  const p0Thread = codexThread({
+    id: "thread-p0",
+    path: "src/auth.ts",
+    line: 24,
+    body: "P0: Keep the authorization bypass blocked.",
+  });
+
+  assert.deepEqual(buildCodexConnectorPolicyBlockDiagnostic(config, [p1Thread, p0Thread]), {
+    count: 2,
+    severity: "P0",
+    file: "src/auth.ts",
+    line: "24",
+    threadUrl: "https://example.test/pr/44#discussion_r1",
+    nextAction: "fix_on_new_head_or_wait_for_github_thread_resolution_or_use_explicit_manual_operator_path",
+  });
+  assert.equal(
+    evaluateCodexConnectorConvergencePolicy(config, { configuredBotCurrentHeadObservedAt: "2026-03-11T00:04:00Z" }, [
+      p0Thread,
+    ])?.outcome,
+    "must_fix_remaining",
+  );
+});
+
+test("Codex Connector policy clusters repeated configured-bot findings", () => {
+  const body =
+    "P1: Missing verifier coverage lets failed restore writes leave a half-restored durable state. Add a regression.";
+  const firstThread = codexThread({
+    id: "thread-restore",
+    path: "src/restore.ts",
+    line: 42,
+    body,
+  });
+  const secondThread = codexThread({
+    id: "thread-restore-test",
+    path: "src/restore.test.ts",
+    line: 88,
+    body,
+  });
+  const unrelatedThread = codexThread({
+    id: "thread-export",
+    path: "src/export.ts",
+    line: 12,
+    body: "P2: Export readiness must reject mixed-snapshot rows instead of stitching partial results together.",
+  });
+
+  const clusters = clusterConfiguredBotReviewThreads([firstThread, secondThread, unrelatedThread]);
+
+  assert.equal(clusters.length, 2);
+  assert.deepEqual(clusters[0]?.threads.map((thread) => thread.id), ["thread-restore", "thread-restore-test"]);
+  assert.deepEqual(clusters[1]?.threads.map((thread) => thread.id), ["thread-export"]);
+});
